### 2020-05-28 #2267 @danielskatz
👋 @RainierBarrett - can you add a little more context about why you would use this software in the README?  The paper seems to have this context, but the README doesn't.
