### 2020-06-01 #2267 @RainierBarrett
Hi guys, thanks for the quick turn around. I added a brief blurb in the README about the main benefits of HOOMD-TF. Let me know what you think.
