### 2020-06-07 #2267 @RMeli
@richardjgowers Looks interesting, count me in.
