### 2020-06-18 #2267 @richardjgowers
Yeah I was waiting to see if one of the earlier invitees with GPU knowledge
wanted to step in but we can proceed with the current reviewers.
