### 2020-06-19 #2367 @whedon
[REVIEW]: HOOMD-TF: GPU-Accelerated, Online Machine Learning in the HOOMD-blue Molecular Dynamics Engine
**Submitting author:** @RainierBarrett (<a href="http://orcid.org/0000-0002-5728-9074">Rainier Barrett</a>)
**Repository:** <a href="https://github.com/ur-whitelab/hoomd-tf/" target ="_blank">https://github.com/ur-whitelab/hoomd-tf/</a>
**Version:** v1.0
