### 2020-06-20 #2367 @RMeli
This is not seem to qualify as a [COI](https://github.com/openjournals/joss/blob/master/COI.md) but  I'll mention it here anyway, for transparency: as @richardjgowers is aware, I am a [MDAnalysis](https://github.com/MDAnalysis/mdanalysis) contributor.

MDAnalysis is used in HOOMD-TF to work with trajectories produced by MD engines other than HOOMD-blue. Despite this, I think I am able to make an impartial assessment of the work.
