### 2020-06-22 f9fee20
fixed issue with tfcompute import
