+++ htf/test-py/test_tensorflow.py
@@ -15,7 +15,7 @@ import math
 import tensorflow as tf
 import build_examples
 
-from hoomd.htf.tfcompute import _make_reverse_indices
+from hoomd.htf.tensorflowcompute import _make_reverse_indices
 
 
 def compute_forces(system, rcut):
@@ -45,7 +45,7 @@ class test_access(unittest.TestCase):
 
     def test_access(self):
         model_dir = build_examples.simple_potential(self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(model_dir,
+        with hoomd.htf.tfcompute(model_dir,
                                  _mock_mode=True) as tfcompute:
             hoomd.context.initialize()
             rcut = 3
@@ -80,7 +80,7 @@ class test_compute(unittest.TestCase):
 
     def test_force_overwrite(self):
         model_dir = build_examples.simple_potential(self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(model_dir) as tfcompute:
+        with hoomd.htf.tfcompute(model_dir) as tfcompute:
             hoomd.context.initialize()
             N = 3 * 3
             NN = N - 1
@@ -106,7 +106,7 @@ class test_compute(unittest.TestCase):
 
     def test_nonlist(self):
         model_dir = build_examples.benchmark_nonlist_graph(self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(model_dir) as tfcompute:
+        with hoomd.htf.tfcompute(model_dir) as tfcompute:
             hoomd.context.initialize()
             system = hoomd.init.create_lattice(
                 unitcell=hoomd.lattice.sq(a=4.0),
@@ -120,7 +120,7 @@ class test_compute(unittest.TestCase):
     def test_full_batch(self):
         hoomd.context.initialize()
         model_dir = build_examples.benchmark_nonlist_graph(self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(model_dir) as tfcompute:
+        with hoomd.htf.tfcompute(model_dir) as tfcompute:
             system = hoomd.init.create_lattice(unitcell=hoomd.lattice.sq(a=4.0),
                                                n=[32, 32])
             hoomd.md.integrate.mode_standard(dt=0.005)
@@ -131,7 +131,7 @@ class test_compute(unittest.TestCase):
     def test_write_empty_tensorboard(self):
         hoomd.context.initialize()
         model_dir = build_examples.benchmark_nonlist_graph(self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(model_dir, write_tensorboard=True) as tfcompute:
+        with hoomd.htf.tfcompute(model_dir, write_tensorboard=True) as tfcompute:
             system = hoomd.init.create_lattice(unitcell=hoomd.lattice.sq(a=4.0),
                                                n=[32, 32])
             hoomd.md.integrate.mode_standard(dt=0.005)
@@ -142,7 +142,7 @@ class test_compute(unittest.TestCase):
 
     def test_trainable(self):
         model_dir = build_examples.trainable_graph(9 - 1, self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(model_dir,
+        with hoomd.htf.tfcompute(model_dir,
                                  write_tensorboard=True) as tfcompute:
             hoomd.context.initialize()
             rcut = 5.0
@@ -161,7 +161,7 @@ class test_compute(unittest.TestCase):
 
     def test_bootstrap(self):
         model_dir = build_examples.trainable_graph(9 - 1, self.tmp)
-        with hoomd.htf.tfcompute.tfcompute(
+        with hoomd.htf.tfcompute(
[…]
