### 2020-06-25 ur-whitelab/hoomd-tf#184 @RMeli
[JOSS] Installation
Hello! I'll open some issues as part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/2367) as I go along. This concerns the installation.

The installation process is rather painless, which is great. I have only a few comments:

* [Installation instructions for BlueHive](https://hoomd-tf.readthedocs.io/en/latest/compilation.html#bluehive-installation) are the first item of the [installation guide](https://hoomd-tf.readthedocs.io/en/latest/installation.html). I think it's worth mentioning what BlueHive is and move the the installation instructions at the end (maybe under a new section "Installing on HPC Clusters"?) in order to avoid confusion for external users.
* The following, under [simple compiling](https://hoomd-tf.readthedocs.io/en/latest/compilation.html#simple-compiling), is a bit misleading: "This method assumes you already have installed hoomd-blue and tensorflow. You could do that, for example, via `conda install -c conda-forge hoomd==2.5.2.`". I suggest to remove the last phrase or add also the command to install TensorFlow and `tbb-devel` (see next item).
* I think a clearer note on the need of `tbb-devel` instead of `tbb` would be useful, since `hoomd` installs `tbb` (and not `tbb-devel`) via conda.
* There are many`-Wsign-compare` and some `-Wreorder` warnings when compiling. It would be nice to get rid of them (discussed in #130)

I also wanted to suggest that a `conda` release (as for HOOMD-blue) will possibly attract a larger user base, but I see that this will be addressed in the future (#96).

---

 [JOSS review](https://github.com/openjournals/joss-reviews/issues/2367)
