### 2020-06-25 ur-whitelab/hoomd-tf#184 @RMeli
[Unit tests](https://hoomd-tf.readthedocs.io/en/latest/installation.html#unit-tests) have a spurious backslash. I think it should be
```
pytest htf/test-py
```
