### 2020-06-25 ur-whitelab/hoomd-tf#185 @RMeli
[JOSS] Quickstart
The [quickstart tutorial](https://hoomd-tf.readthedocs.io/en/latest/quickstart.html) throws the following errors:

* `NameError: name 'energy' is not defined`
* `NameError: name 'hoomd' is not defined`

`import hoomd` is missing, and `energy` is not defined (`particle_energy` instead?).

---

[JOSS review](https://github.com/openjournals/joss-reviews/issues/2367)
