### 2020-06-25 12460f7
Updated installation and unit test to address JOSS review
