### 2020-06-25 a4f6387
Unit test fixes (#186)
* Fixed imports and position dependence issues

* Removed npz file from tests
