### 2020-06-25 0fc61da
Added note about sanity check for installs
