### 2020-06-25 0473eeb
Addressed all compiler warnings
