+++ htf/tf2hoomd_op/tf2hoomd.cc
@@ -50,7 +50,7 @@ class TfToHoomdOp : public OpKernel {
     const Tensor& input_tensor = context->input(0);
     auto input = input_tensor.flat<T>();
 
-    if (input.size() > m_output_memory->mem_size / m_output_memory->element_size ) {
+    if (static_cast<unsigned int> (input.size()) > m_output_memory->mem_size / m_output_memory->element_size ) {
       errors::InvalidArgument(
           "Tensor input size is too large for output buffer!");
     }
