### 2020-06-25 1815816
JOSS Install Documentation Updates (#187)
* Updated installation and unit test to address JOSS review

* Added note about sanity check for installs
