### 2020-06-25 ur-whitelab/hoomd-tf#185 @whitead
Thanks for doing our review! We've addressed this one with #188. Please close if satisfied or let me know if you have other concerns.
