### 2020-06-25 ur-whitelab/hoomd-tf#184 @whitead
Hi @RMeli! Thank you for doing our review and we appreciate your helpful suggestions! We've added the following PRs to address your notes:

1. #187 revised the documentation based on your notes on `tbb-devel`, the simple compiling, and our HPC install
2.  #189 Addresses all compiler warnings
3. #186 Fixes unit tests, aside from the one you mentioned (MDAnalysis), which we'll address via #190
