### 2020-06-25 ur-whitelab/hoomd-tf#184 @whitead
Also, please refer to this [doc website](https://hoomd-tf.readthedocs.io/en/readthedocs/). We'll update the `latest` version once we're done and have bumped the version.
