### 2020-06-25 ur-whitelab/hoomd-tf#185 @RMeli
LGTM, thanks for addressing this!
