### 2020-06-25 ur-whitelab/hoomd-tf#184 @RMeli
Thanks, LGTM. I'll close this, given #190.

> Also, please refer to this doc website. We'll update the latest version once we're done and have bumped the version.

Sure, not problem!
