### 2020-06-25 8e7ec0f
Addressed all compiler warnings (#189)
