### 2020-06-26 ur-whitelab/hoomd-tf#184 @malramsay64
Thanks @RMeli for opening this issue, I was also going to suggest similar changes. @whitead the install instructions are much clearer particularly for the simple case.
