### 2020-06-26 ur-whitelab/hoomd-tf#184 @RMeli
@malramsay64 If you have other suggestions feel free to post them here and I'll re-open the issue (in order not to clutter the issue tracker with too many issues for the JOSS review).
