### 2020-06-26 ur-whitelab/hoomd-tf#192 @RMeli
[JOSS] Paper
I think the manuscript is good. I only have a few comments about the references:

* Reference #1 (Abadi et al., 2015) contains spurious `~`
* Reference #5 (Eastman, 2018) is missing a link? It should probably link to [openmm-nn](https://github.com/openmm/openmm-nn)
* Some reference that should be part of the text are still appearing within parenthesis (`For example (Rupp, Tkatchenko, Müller, & Anatole Von Lilienfeld, 2012) and (Botu & Ramprasad, 2015) are benchmark works...`). I'm not sure if there is a way to remove them...

I'm also wondering if a few more reference should be included, especially about force-matching, force-field biasing/learning and collective variables?

---
[JOSS review](https://github.com/openjournals/joss-reviews/issues/2367)
