### 2020-06-26 c735c0f
Update JOSS paper with more refs, housekeeping
