+++ paper.bib
@@ -156,3 +156,105 @@ url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
 volume = {32},
 year = {2011}
 }
+@misc{WangMLSampling2020,
+abstract = {Molecular dynamics (MD) has become a powerful tool for studying biophysical systems, due to increasing computational power and availability of software. Although MD has made many contributions to better understanding these complex biophysical systems, there remain methodological difficulties to be surmounted. First, how to make the deluge of data generated in running even a microsecond long MD simulation human comprehensible. Second, how to efficiently sample the underlying free energy surface and kinetics. In this short perspective, we summarize machine learning based ideas that are solving both of these limitations, with a focus on their key theoretical underpinnings and remaining challenges.},
+archivePrefix = {arXiv},
+arxivId = {1909.11748},
+author = {Wang, Yihang and {Lamim Ribeiro}, Jo{\~{a}}o Marcelo and Tiwary, Pratyush},
+booktitle = {Current Opinion in Structural Biology},
+doi = {10.1016/j.sbi.2019.12.016},
+eprint = {1909.11748},
+file = {::},
+issn = {1879033X},
+month = {apr},
+pages = {139--145},
+pmid = {31972477},
+publisher = {Elsevier Ltd},
+title = {{Machine learning approaches for analyzing and enhancing molecular dynamics simulations}},
+volume = {61},
+year = {2020}
+}
+@article{TraplAnnColvar2019,
+abstract = {The state of a molecular system can be described in terms of collective variables. These low-dimensional descriptors of molecular structure can be used to monitor the state of the simulation, to calculate free energy profiles or to accelerate rare events by a bias potential or a bias force. Frequent calculation of some complex collective variables may slow down the simulation or analysis of trajectories. Moreover, many collective variables cannot be explicitly calculated for newly sampled structures. In order to address this problem, we developed a new package called anncolvar. This package makes it possible to build and train an artificial neural network model that approximates a collective variable. It can be used to generate an input for the open-source enhanced sampling simulation PLUMED package, so the collective variable can be monitored and biased by methods available in this program. The computational efficiency and the accuracy of anncolvar are demonstrated on selected molecular systems (cyclooctane derivative, Trp-cage miniprotein) and selected collective variables (Isomap, molecular surface area).},
+author = {Trapl, Dalibor and Horvacanin, Izabela and Mareska, Vaclav and Ozcelik, Furkan and Unal, Gozde and Spiwok, Vojtech},
+doi = {10.3389/fmolb.2019.00025},
+file = {::},
+issn = {2296-889X},
+journal = {Frontiers in Molecular Biosciences},
+keywords = {Collective variables,Free energy simulations,Metadynamics,Molecular dynamics simulation,Neural networks},
+month = {apr},
+number = {APR},
+pages = {25},
+publisher = {Frontiers Media S.A.},
+title = {{Anncolvar: Approximation of Complex Collective Variables by Artificial Neural Networks for Analysis and Biasing of Molecular Simulations}},
+url = {https://www.frontiersin.org/article/10.3389/fmolb.2019.00025/full},
+volume = {6},
+year = {2019}
+}
+@article{SmithANI2017,
[…]
