### 2020-06-26 ur-whitelab/hoomd-tf#192 @RainierBarrett
Thanks for your review. I'm not really familiar with doing these citations in markdown, so I'm not sure about the parentheses either. I pushed an updated paper/bib to the joss-paper-branch to address the rest of these points. Let us know if you have any other comments!
