### 2020-06-26 ur-whitelab/hoomd-tf#192 @RMeli
Thanks @RainierBarrett . I'll investigate how to deal with references within the text, but it seems that the current changes broke the references. I think the square brackets `[]` were needed?

See the problem here:
> [ 👉 Check article proof 📄 👈 ](https://github.com/openjournals/joss-papers/blob/joss.02367/joss.02367/10.21105.joss.02367.pdf)
