### 2020-06-26 ur-whitelab/hoomd-tf#192 @RMeli
According to the [Example paper and bibliography](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) it looks like `@reference` prints the reference within the text while `[@author:2001]` prints the reference within parentheses (if I understand correctly).  I think the `` are just there to render the example as code? It's a bit confusing...
