### 2020-06-27 9d13412
Added checks for neighbor list overflow and prevented GPU segfaults
