### 2020-06-27 85914d8
Added docs for new check_nlist flag
