+++ htf/graphbuilder.py
@@ -14,7 +14,7 @@ class graph_builder:
         :param output_forces: True if your graph will compute forces to be used in TensorFlow
         :type output_forces: bool
         :type check_nlist: bool
-        :param check_nlist: True will raise error if neighbor 
+        :param check_nlist: True will raise error if neighbor
                             list overflows (nneighbor_cutoff too low)
     """
 
@@ -59,12 +59,12 @@ class graph_builder:
 
         # add check for nlist size
         if check_nlist:
-            NN = tf.reduce_max(tf.reduce_sum(tf.cast(self.nlist[:, :, 0] > 0, 
-                                                     tf.dtypes.int32), axis=1), 
-                                                     axis=0)
+            NN = tf.reduce_max(tf.reduce_sum(tf.cast(self.nlist[:, :, 0] > 0,
+                                                     tf.dtypes.int32), axis=1),
+                               axis=0)
             check_op = tf.Assert(tf.less(NN, nneighbor_cutoff), ['Neighbor list is full!'])
             self.out_nodes.append(check_op)
-        
+
 
     ## \var atom_number
     # \internal
