### 2020-06-29 ur-whitelab/hoomd-tf#192 @RainierBarrett
@RMeli You're talking about this section, right?

```
For a quick reference, the following citation commands can be used:
- `@author:2001`  ->  "Author et al. (2001)"
- `[@author:2001]` -> "(Author et al., 2001)"
- `[@author1:2001; @author2:2001]` -> "(Author1 et al., 2001; Author2 et al., 2002)"
```

That's strange, because all of our refs are in the style of the first bullet there, but I think they were showing up as parentheticals the first time... I'll play around with it and see what happens.
