+++ paper.bib
@@ -156,6 +156,7 @@ url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/jcc.21787},
 volume = {32},
 year = {2011}
 }
+
 @misc{WangMLSampling2020,
 abstract = {Molecular dynamics (MD) has become a powerful tool for studying biophysical systems, due to increasing computational power and availability of software. Although MD has made many contributions to better understanding these complex biophysical systems, there remain methodological difficulties to be surmounted. First, how to make the deluge of data generated in running even a microsecond long MD simulation human comprehensible. Second, how to efficiently sample the underlying free energy surface and kinetics. In this short perspective, we summarize machine learning based ideas that are solving both of these limitations, with a focus on their key theoretical underpinnings and remaining challenges.},
 archivePrefix = {arXiv},
@@ -174,6 +175,7 @@ title = {{Machine learning approaches for analyzing and enhancing molecular dyna
 volume = {61},
 year = {2020}
 }
+
 @article{TraplAnnColvar2019,
 abstract = {The state of a molecular system can be described in terms of collective variables. These low-dimensional descriptors of molecular structure can be used to monitor the state of the simulation, to calculate free energy profiles or to accelerate rare events by a bias potential or a bias force. Frequent calculation of some complex collective variables may slow down the simulation or analysis of trajectories. Moreover, many collective variables cannot be explicitly calculated for newly sampled structures. In order to address this problem, we developed a new package called anncolvar. This package makes it possible to build and train an artificial neural network model that approximates a collective variable. It can be used to generate an input for the open-source enhanced sampling simulation PLUMED package, so the collective variable can be monitored and biased by methods available in this program. The computational efficiency and the accuracy of anncolvar are demonstrated on selected molecular systems (cyclooctane derivative, Trp-cage miniprotein) and selected collective variables (Isomap, molecular surface area).},
 author = {Trapl, Dalibor and Horvacanin, Izabela and Mareska, Vaclav and Ozcelik, Furkan and Unal, Gozde and Spiwok, Vojtech},
@@ -191,6 +193,7 @@ url = {https://www.frontiersin.org/article/10.3389/fmolb.2019.00025/full},
 volume = {6},
 year = {2019}
 }
+
 @article{SmithANI2017,
 abstract = {Deep learning is revolutionizing many areas of science and technology, especially image, text, and speech recognition. In this paper, we demonstrate how a deep neural network (NN) trained on quantum mechanical (QM) DFT calculations can learn an accurate and transferable potential for organic molecules. We introduce ANAKIN-ME (Accurate NeurAl networK engINe for Molecular Energies) or ANI for short. ANI is a new method designed with the intent of developing transferable neural network potentials that utilize a highly-modified version of the Behler and Parrinello symmetry functions to build single-atom atomic environment vectors (AEV) as a molecular representation. AEVs provide the ability to train neural networks to data that spans both configurational and conformational space, a feat not previously accomplished on this scale. We utilized ANI to build a potential called ANI-1, which was trained on a subset of the GDB databases with up to 8 heavy atoms in order to predict total energies for organic molecules containing four atom types: H, C, N, and O. To obtain an accelerated but physically relevant sampling of molecular potential surfaces, we also proposed a Normal Mode Sampling (NMS) method for generating molecular conformations. Through a series of case studies, we show that ANI-1 is chemically accurate compared to reference DFT calculations on much larger molecular systems (up to 54 atoms) than those included in the training data set.},
 archivePrefix = {arXiv},
@@ -210,6 +213,7 @@ url = {https://pubs.rsc.org/en/content/articlehtml/2017/sc/c6sc05720a https://pu
 volume = {8},
[…]
