### 2020-06-29 ur-whitelab/hoomd-tf#192 @RMeli
@RainierBarrett yes, that's the section. If I'm not mistaken, before you had [@author:2001] (not \`[@author:2001]\`) and it was working. So I assume the \`\` are spurious (i.e. they were supposed to render to a code block in the documentation?)... I'd try using `@author:2001` for a reference within the text and `[@author:2001]` for a reference within parentheses (without ``).
