+++ paper.bib
@@ -264,3 +264,10 @@ title = {{TorchANI: A Free and Open Source PyTorch Based Deep Learning Implement
 url = {https://pubs.acs.org/doi/10.1021/acs.jcim.0c00451},
 year = {2020}
 }
+
+@misc{PyTorch,
+   author = {Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory},
+   year = {2016},
+   note = {https://github.com/pytorch/pytorch},
+   title = {PyTorch}
+}
