+++ paper.md
@@ -38,22 +38,22 @@ For example, one can run a molecular dynamics (MD) or density functional theory 
 input data to train a deep neural network that reproduces the energetics at a fraction of the computational cost.
 Some recent examples include an energy-conserving force-field learned with a custom gradient-domain
 model [@ChmielaConservedEnergyMLFF2017], DFT-based neural network force-fields [@SmithDFTNNPotential2017],
-and a neural network coarse-grained potential @[WangCGML2019]. Other applications include the use of ML methods for collective variable (CV) calculation @[TraplAnnColvar2019] and enhanced sampling @[WangMLSampling2020] for MD simulations. 
-A limitation of these methods is that they treat the molecular modeling calculations as a static dataset, whereas molecular modeling calculations can be treated as interrogative functions, opening the door to methods like reinforcement learning or active learning. Thus there is a clear limitation of existing implementations due to their sequential nature of going from calculation to ML. Another practical issue is that neural network force-field implementations often duplicate standard ML frameworks, preventing them from keeping progress with state-of-the art methods. For example @[RuppMLAtomizationE2012] and @[BotuMLQMMD] are benchmark works in this field and custom implementations. This limits the scope and speed of translating ML advances to the molecular modeling community.
+and a neural network coarse-grained potential [@WangCGML2019]. Other applications include the use of ML methods for collective variable (CV) calculation [@TraplAnnColvar2019] and enhanced sampling [@WangMLSampling2020] for MD simulations. 
+A limitation of these methods is that they treat the molecular modeling calculations as a static dataset, whereas molecular modeling calculations can be treated as interrogative functions, opening the door to methods like reinforcement learning or active learning. Thus there is a clear limitation of existing implementations due to their sequential nature of going from calculation to ML. Another practical issue is that neural network force-field implementations often duplicate standard ML frameworks, preventing them from keeping progress with state-of-the art methods. For example @RuppMLAtomizationE2012 and @BotuMLQMMD are benchmark works in this field and custom implementations. This limits the scope and speed of translating ML advances to the molecular modeling community.
 
 
 This need has led us to develop HOOMD-TF, a flexible direct integration of a standard ML library and standard molecular simulation framework that maintains GPU acceleration. Our goals are to improve the reproducibility of ML methods in molecular simulation, ease translation of ML advances, and remove the need for sequential simulation and ML. This should enable active learning, reinforcement learning, and online learning of molecular simulations. 
 
-There are other applications focusing on ML in molecular modeling, such as [DeepChem](https://www.deepchem.io/) and [@Aspuru-GuzikMaterialsDiscovery2015],
+There are other applications focusing on ML in molecular modeling, such as [DeepChem](https://www.deepchem.io/) and @Aspuru-GuzikMaterialsDiscovery2015,
 which are largely concerned with property prediction and representation. There are also similar works to HOOMD-TF called
-[OpenMM-NN](https://github.com/openmm/openmm-nn) @[EastmanOpenMMNN2018], which allows the use of pre-trained TensorFlow models in [OpenMM](http://openmm.org/) @[PandeOpenMM2013], and TorchANI @[GaoTorchANI2020], which uses PyTorch @[PyTorch] for similar purposes. In contrast,
+[OpenMM-NN](https://github.com/openmm/openmm-nn) [@EastmanOpenMMNN2018], which allows the use of pre-trained TensorFlow models in [OpenMM](http://openmm.org/) [@PandeOpenMM2013], and TorchANI [@GaoTorchANI2020], which uses PyTorch [@PyTorch] for similar purposes. In contrast,
[…]
