+++ paper.bib
@@ -268,6 +268,6 @@ year = {2020}
 @misc{PyTorch,
    author = {Paszke, Adam and Gross, Sam and Chintala, Soumith and Chanan, Gregory},
    year = {2016},
-   note = {https://github.com/pytorch/pytorch},
+   url = {https://github.com/pytorch/pytorch},
    title = {PyTorch}
 }
