### 2020-06-29 ur-whitelab/hoomd-tf#192 @RainierBarrett
Okay, yeah I was just confused because of the formatting thing going on in the example paper there. Latest push should address all these. Thanks!
