### 2020-06-29 ur-whitelab/hoomd-tf#192 @RMeli
I agree it's quite confusing... =)

LGTM now, thanks for addressing the comments. I'll leave this open in case @malramsay64 has further comments.
