### 2020-06-29 fc5e03e
Added check on type for save_tensor
