### 2020-06-29 #2367 @richardjgowers
Hi @RainierBarrett has [this preprint](https://chemrxiv.org/articles/A_GPU-Accelerated_Machine_Learning_Framework_for_Molecular_Simulation_Hoomd-Blue_with_TensorFlow/8019527) been submitted elsewhere?
