### 2020-06-29 e378931
Added check on type for save_tensor (#194)
Added check on type for save_tensor
