### 2020-06-30 be8b299
Added checks for neighbor list overflow and prevented GPU segfaults (#193)
* Added checks for neighbor list overflow and prevented GPU segfaults

* Added docs for new check_nlist flag
