### 2020-06-30 ur-whitelab/hoomd-tf#184 @malramsay64
@RMeli I have a couple more points for the installation since I have found a range of situations where the installation could use some clarification.

[ ] gcc 7 is not listed as a dependency. With the major version appearing to be required for compilation, to match the version used to compile tensorflow
[ ] There is no list of supported python versions. `pip install tensorflow==1.15` doesn't work for python 3.8.
[ ] In the instructions for running the tests, it should be a noted that `pytest` and `MDAnalysis` need to be installed to run all the tests.

Also not strictly required, but it might help readability to include the code snippet for the installation of tensorflow in the [simple installation step](https://hoomd-tf.readthedocs.io/en/readthedocs/compilation.html#simple-compiling) as you have done for hoomd.
