### 2020-06-30 ur-whitelab/hoomd-tf#192 @malramsay64
The paper looks good to me as well
