### 2020-06-30 1bec947
Added note on gcc versions
