### 2020-06-30 ur-whitelab/hoomd-tf#184 @whitead
Hi @malramsay64, thanks for doing this review. I've updated the [documentation](https://hoomd-tf.readthedocs.io/en/readthedocs/compilation.html) with your feedback. Specifically:

  1. gcc 7 is not required, just the compiler used to compile tensorflow must be available on the machine. I added a note about this.
  2. Same for python versions, added a note on this too.
  3. `pytest` note added. The `MDAnalysis` was noted by @RMeli and we'll get it done with #190.

Thanks and let me know if you have any other issues.

I put the code snippet in the simple compiling section as suggested.
