### 2020-07-01 ur-whitelab/hoomd-tf#184 @malramsay64
@whitead LGTM

@RMeli I have no further points so happy for this issue to be closed.
