### 2020-07-02 ed94c86
conditional imports for trajectory parsing (#196)
Conditional statement with skiptest added

* utilities.rst,  compilation.rst updated

Docs added for trajectory parsing

Optional Dependencies added under Installing HOOMD-TF

Issue #190 addressed
