### 2020-07-02 #2367 @RainierBarrett
@richardjgowers I'm in the midst of wrapping up my PhD thesis right now, sorry for the delayed reply. This pre-print is not submitted or under review anywhere presently, and this JOSS submission is the only manuscript we have describing the HOOMD-TF package. We had previously submitted the content in that pre-print to a comp chem journal, but as you may have guessed they didn't like a software-only paper so we've arrived here. We're debating what to do about the chemrxiv pre-print, perhaps linking here as the final version should our article be accepted at JOSS.
