### 2020-07-03 ur-whitelab/hoomd-tf#197 @whitead
Thanks for pointing out these issues. Some of these notebooks are indeed out of date and not that descriptive. Hopefully I've addressed this in PR #198
