### 2020-07-06 ur-whitelab/hoomd-tf#199 @RMeli
[JOSS] Community Guidelines
Can you please add some [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) to the repository?

---
[JOSS review](https://github.com/openjournals/joss-reviews/issues/2367)
