### 2020-07-06 ur-whitelab/hoomd-tf#200 @RMeli
[JOSS] Documentation and Readme
The description on the documentation homepage is quite short (see [statement of need](https://joss.readthedocs.io/en/latest/review_criteria.html#a-statement-of-need)). Since the README contains a more detailed explanation, I think using the same explanation in both the README and the documentation would be beneficial.

Listing the main (and optional) dependencies mentioned in the documentation on the README would be helpful as well.

---
[JOSS review](https://github.com/openjournals/joss-reviews/issues/2367)
