### 2020-07-08 ur-whitelab/hoomd-tf#201 @malramsay64
[JOSS] Tests and Benchmarks
There are a couple of currently undocumented files/functionality that are hiding away in the source code that I would like some clarification on.

- [x] The [test-cc directory](https://github.com/ur-whitelab/hoomd-tf/tree/master/htf/test-cc) that appears to contain some element of testing the C++ components. I can see it was intentional to create (#9 #48), so it looks to be an omission that there is no documentation of this test.

- [x] The [benchmarking directory](https://github.com/ur-whitelab/hoomd-tf/tree/master/htf/models/benchmarking) that contains a performance benchmark. There is definitely benefit to having a performance benchmark as a method of ensuring the appropriate hardware is being used.
---
[JOSS Review](https://github.com/openjournals/joss-reviews/issues/2367)
