### 2020-07-08 ur-whitelab/hoomd-tf#197 @malramsay64
Changes are great. LGTM
