### 2020-07-08 5a749e9
Add more detailed description from README to docs homepage
