### 2020-07-08 ur-whitelab/hoomd-tf#199 @RainierBarrett
PR #203 should address this. I'm PRing these into the readthedocs branch for now so we can get that included in the version bump with everything else once that's ready to go.
