### 2020-07-08 ur-whitelab/hoomd-tf#200 @RainierBarrett
PR #202 should cover this. I'm PRing these into the readthedocs branch for now so we can get that included in the version bump with everything else once that's ready to go.
