### 2020-07-08 ur-whitelab/hoomd-tf#200 @RMeli
Sounds good, thanks! LGTM
