### 2020-07-08 ur-whitelab/hoomd-tf#199 @RMeli
Thanks, the code of conduct looks good! Can you please add some contribution guidelines as well? I think a few lines linking to the Issue and PR tabs are more than enough (unless you also have a mailing list/Slack channel/...)

> There should be clear guidelines for third-parties wishing to:
>
>  *  Contribute to the software
>  * Report issues or problems with the software
>  *  Seek support
