### 2020-07-08 ur-whitelab/hoomd-tf#201 @RainierBarrett
Hi @malramsay64. The test-cc directory is a remnant from early development. We have since moved to python testing for this functionality (see `test_compute` in test_tensorflow.py), so this should probably just be removed. Thanks for the catch. I'll work on adding some documentation around benchmarking.
