### 2020-07-08 986cbc6
Add instructions for PRs and issues
