### 2020-07-09 b233ca6
added a tutorial and mapping function to utils.
