### 2020-07-09 ur-whitelab/hoomd-tf#200 @RMeli
I'll close this unless @malramsay64 has anything to add.
