### 2020-07-09 ur-whitelab/hoomd-tf#199 @RMeli
Commit 986cbc6 solves this issue for me, thanks. I'll close it unless @malramsay64 has any additional comments.
