### 2020-07-09 92b65a9
Remove obsolete c++ test
