### 2020-07-09 147ebf6
Issue 199 (#203)
* Add code of conduct

* Add instructions for PRs and issues
