### 2020-07-09 78a9c86
Added test for size of neighborlists
