### 2020-07-09 1af3d31
Added test for size of neighborlists
