### 2020-07-09 bd4ece1
Made it possible pass mass-weighted matrix
