### 2020-07-09 90115c9
Update HOOMD ref to 2020
