+++ paper.bib
@@ -271,3 +271,18 @@ year = {2020}
    url = {https://github.com/pytorch/pytorch},
    title = {PyTorch}
 }
+@article{Anderson2020HOOMD,
+archivePrefix = {arXiv},
+arxivId = {1308.5587},
+author = {Anderson, Joshua A. and Glaser, Jens and Glotzer, Sharon C.},
+doi = {10.1016/j.commatsci.2019.109363},
+eprint = {1308.5587},
+issn = {09270256},
+journal = {Computational Materials Science},
+keywords = {CUDA,GPU,Molecular dynamics,Molecular simulation,Monte Carlo,Python},
+pages = {109363},
+publisher = {Elsevier B.V.},
+title = {{HOOMD-blue: A Python package for high-performance molecular dynamics and hard particle Monte Carlo simulations}},
+volume = {173},
+year = {2020}
+}
