### 2020-07-09 a9f2e0b
Remove superfluous directory, rename benchmark script
