### 2020-07-09 0bb77ce
Update benchmarking script, runs on cpu and gpu
