### 2020-07-10 049960f
Add benchmarking to docs. Fix minor syntax in unit tests
