+++ sphinx-docs/source/unit_tests.rst
@@ -3,8 +3,8 @@
 Unit Tests
 ----------
 
-You can run the unit tests directly via `python htf/test-py/test_tensorflow.py`,
-`python htf/test-py/test_utils.py`, etc. For convenience, you can also install the `pyest` package and run the following from the root
+You can run the unit tests directly via ``python htf/test-py/test_tensorflow.py``,
+``python htf/test-py/test_utils.py``, etc. For convenience, you can also install the ``pyest`` package and run the following from the root
 HOOMD-TF directory:
 
 .. code:: bash
