### 2020-07-10 3628b08
Add benchmarking to docs. Fix minor syntax in unit tests
