### 2020-07-10 d25fdf5
Fix pep8 compliance
