### 2020-07-10 ur-whitelab/hoomd-tf#201 @RainierBarrett
Okay, #207 should address this. I don't know how but my git bash on windows got a little antsy so there's some wonky commit history going on.
