### 2020-07-11 #2367 @RMeli
@richardjgowers , I think only https://github.com/ur-whitelab/hoomd-tf/issues/201 is outstanding which should be addressed by https://github.com/ur-whitelab/hoomd-tf/pull/208.
