### 2020-07-11 ur-whitelab/hoomd-tf#201 @RMeli
Addressed in #208 (superseding #207)
