### 2020-07-16 3d1026d
Issue 201 fix (#208)
* Remove obsolete c++ test

* Update benchmarking script, runs on cpu and gpu

* Add benchmarking to docs.
