### 2020-07-16 fbcb9b6
Readthedocs (#209)
* Initial setup for readthedocs

* Remove python keyword from config yaml

* Remove conda env after docs change

* Update README with new docs link

* Updated install docs

* Added note on gcc versions

* Add dependencies to README

* Add code of conduct

* Add instructions for PRs and issues

Co-authored-by: Andrew White <white.d.andrew@gmail.com>
