+++ sphinx-docs/source/unit_tests.rst
@@ -1,7 +1,8 @@
 .. _unit_tests:
-   
+
 Unit Tests
 ----------
+
 
 You can run the unit tests directly via ``python htf/test-py/test_tensorflow.py``,
 ``python htf/test-py/test_utils.py``, etc. For convenience, you can also install the ``pyest`` package and run the following from the root
@@ -10,4 +11,4 @@ HOOMD-TF directory:
 .. code:: bash
 
     pytest htf/test-py/
-   
+
