### 2020-07-16 66175df
changed beads distribution to a list input
