### 2020-07-17 99789bb
Added changelog for latest release and JOSS icon
