### 2020-07-17 39f532c
Added JOSS info to sphinx index
