+++ sphinx-docs/source/index.rst
@@ -22,7 +22,8 @@ HOOMD-TF includes several utility functions as convenient built-ins, such as:
 
 In addition to all these, the TensorFlow interface of HOOMD-TF makes implementing arbitrary ML models as easy as it is in TensorFlow, by exposing the HOOMD-blue neighbor list and particle positions to TensorFlow. This enables GPU-accelerated tensor calculations, meaning arbitrary collective variables can be treated in the TensorFlow model framework, as long as they can be expressed as tensor operations on particle positions or neighbor lists.
 
-# Citation
+Citation
+-------
 
 |status|
 
@@ -35,7 +36,8 @@ Please use the following citation:
    HA Gandhi, G Wellawatte, and AD White (2020) *Journal of Open Source Software* doi: 10.21105/joss.02367
 
 
-# Tutorials
+Tutorial
+--------
 
 See `example notebooks here <https://nbviewer.jupyter.org/github/ur-whitelab/hoomd-tf/tree/master/examples/>`_ to learn about what Hoomd-TF can do.
 
