### 2020-07-20 201668f
Added changelog for latest release and JOSS icon (#212)
* Added changelog for latest release and JOSS icon

* RST hyperlinks

* Added JOSS info to sphinx index

* Fixed section headings
