+++ sphinx-docs/source/index.rst
@@ -22,9 +22,24 @@ HOOMD-TF includes several utility functions as convenient built-ins, such as:
 
 In addition to all these, the TensorFlow interface of HOOMD-TF makes implementing arbitrary ML models as easy as it is in TensorFlow, by exposing the HOOMD-blue neighbor list and particle positions to TensorFlow. This enables GPU-accelerated tensor calculations, meaning arbitrary collective variables can be treated in the TensorFlow model framework, as long as they can be expressed as tensor operations on particle positions or neighbor lists.
 
-# Tutorials
+Citation
+-------
 
-See `example notebooks here <https://nbviewer.jupyter.org/github/ur-whitelab/hoomd-tf/tree/master/examples/>`_ to learn about what Hoomd-TF can do. 
+|status|
+
+.. |status| image:: https://joss.theoj.org/papers/5d1323eadec82aabe86c65a403ff8f90/status.svg
+   :target: https://joss.theoj.org/papers/5d1323eadec82aabe86c65a403ff8f90
+
+Please use the following citation:
+
+   HOOMD-TF: GPU-Accelerated, Online Machine Learning in the Hoomd-blue Molecular Dynamics Engine. R Barrett, M Chakraborty, DB Amirkulova,
+   HA Gandhi, G Wellawatte, and AD White (2020) *Journal of Open Source Software* doi: 10.21105/joss.02367
+
+
+Tutorial
+--------
+
+See `example notebooks here <https://nbviewer.jupyter.org/github/ur-whitelab/hoomd-tf/tree/master/examples/>`_ to learn about what Hoomd-TF can do.
 
 
 .. toctree::
@@ -48,5 +63,5 @@ See `example notebooks here <https://nbviewer.jupyter.org/github/ur-whitelab/hoo
    :caption: Python Code
 
    htf_package
-   
+
 * :ref:`search`
