### 2020-07-20 9ec34c0
Added check for skew in boxes and test
