### 2020-07-20 dab3a17
Bumped version to 1.0
