+++ sphinx-docs/source/changelog.rst
@@ -1,7 +1,7 @@
 Change Log
 ==========
 
-v0.7 (7/17/2020)
+v1.0 (7/20/2020)
 ----
 
 *JOSS Review*
