### 2020-07-20 13a2b2c
some typo fixes in docs
