+++ sphinx-docs/source/index.rst
@@ -16,9 +16,10 @@ HOOMD-TF was designed with speed in mind, and minimizes latency with a GPU-GPU c
 HOOMD-TF allows for online machine learning with early termination, rather than the more tradditional batch learning approach for MD+ML.
 
 HOOMD-TF includes several utility functions as convenient built-ins, such as:
-* RDF calculation
-* EDS Biasing (See `this paper <https://www.tandfonline.com/doi/full/10.1080/08927022.2019.1608988>`_)
-* Coarse-Grained simulation force matching
+
+- RDF calculation
+- EDS Biasing (See `this paper <https://www.tandfonline.com/doi/full/10.1080/08927022.2019.1608988>`_)
+- Coarse-Grained simulation force matching
 
 In addition to all these, the TensorFlow interface of HOOMD-TF makes implementing arbitrary ML models as easy as it is in TensorFlow, by exposing the HOOMD-blue neighbor list and particle positions to TensorFlow. This enables GPU-accelerated tensor calculations, meaning arbitrary collective variables can be treated in the TensorFlow model framework, as long as they can be expressed as tensor operations on particle positions or neighbor lists.
 
