### 2020-07-20 #2367 @RainierBarrett
Okay, we've bumped to v1.0 and I set up a zenodo here: https://zenodo.org/record/3952665
