### 2020-07-23 5dfce10
added new method for finding molecules from a gromacs topology. Added example to jupyter notebook.
