### 2020-07-25 #2367 @Kevin-Mattheus-Moerman
@RainierBarrett Here are some additional points to work on for your paper:

- [ ] Please complete your affiliations to include city, state (if applicable), and country (do not use acronyms for states and countries).
- [ ] Please amend the Zenodo archive to have its own independent Zenodo DOI. I believe currently you made it have the same DOI as the (future) JOSS paper. This is not permitted as both should have their own DOI, so you have to change the archived version DOI (@richardjgowers for future reference please check that the archive DOI is accurate/appropriate in this sense) and report back with it here.
