### 2020-07-25 383c371
Add city/state/nation to affiliations
