### 2020-07-25 #2367 @RainierBarrett
@Kevin-Mattheus-Moerman I updated author affiliations in the manuscript on joss-paper-branch. Unfortunately, the zenodo website isn't allowing me to change the DOI to a zenodo-specific one. When I try to ask it to reserve one in the edit process, it gives an error:

> Digital Object Identifier: The prefix 10.5281 is administrated locally.

How shall we proceed?
