### 2020-07-25 #2367 @Kevin-Mattheus-Moerman
@openjournals/dev @arfon do you know more about this. Can you help advise on what to do. The archive accidentally caries the joss DOI reserved for this paper. The author says they cannot alter it but we are still looking into it.

@RainierBarrett I just checked and I clicked edit on one of my archived software projects. It looks like I can alter the DOI and there is a comment that says: `Please note that it is NOT possible to edit a Zenodo DOI once it has been registered by us, while it is always possible to edit a custom DOI.`. You should have the latter case. Admittedly I did not click proceed as I didn't want to change the DOI but perhaps you can check it again yourself.
