### 2020-07-25 #2367 @Kevin-Mattheus-Moerman
> @Kevin-Mattheus-Moerman yeah I saw that as well, so I tried to totally delete the DOI entry and click the "Reserve DOI" button, that's what caused this error. If I just leave the DOI field blank it simply keeps the JOSS one from before.

Okay. Lets wait to see what @arfon @openjournals/dev come back with
