### 2020-07-26 #2367 @arfon
> @Kevin-Mattheus-Moerman yeah I saw that as well, so I tried to totally delete the DOI entry and click the "Reserve DOI" button, that's what caused this error. If I just leave the DOI field blank it simply keeps the JOSS one from before.

I'm afraid you're going to have to ask Zenodo support for help here. The DOI you've reserved is the JOSS DOI that will be used for the paper (and can't resolve to Zenodo). Please ask them to delete this record and allow you to make another one.
