### 2020-07-26 #2367 @Kevin-Mattheus-Moerman
@arfon if deletion is a long and slow process, there is an ugly work around. If the authors minted a new release (annoying but works), they could create a new archive with a different DOI. This way we could move on with this paper (and the other archived version can be removed at a later point). Let's try waiting for deletion first though.
