### 2020-07-26 #2367 @RainierBarrett
Thanks @arfon I'll try asking Zenodo to delete and get back as soon as I hear from them.
