### 2020-07-27 2166a8d
Removed deadlock issue
