### 2020-07-27 #2367 @RainierBarrett
Alright all @arfon @Kevin-Mattheus-Moerman @RMeli @richardjgowers, Zenodo support was quick to reply and we just had to delete the old archive and re-upload to get a specific DOI for the archive. It's up now at https://zenodo.org/record/3962305. Thanks for your patience, and sorry for the confusion.
