### 2020-07-27 b93ec8d
Added release notes for v1.0.1
