### 2020-07-28 #2367 @kthyng
Your paper looks good. 👍
