### 2020-07-28 #2367 @RainierBarrett
Hi @kthyng , since the Zenodo archive was made with 1.0, I guess that makes the most sense for this, too. Any updates since publication should be easy enough to find :)
